fixed stacked model in bash example, manage missing categories

+++ bash_examples/train_stack.py
@@ -1,7 +1,7 @@
 
 
 import pandas as pd
-from dsipts import TimeSeries, beauty_string,extend_time_df
+from dsipts import TimeSeries, beauty_string,extend_time_df,read_public_dataset
 from omegaconf import DictConfig, OmegaConf,ListConfig
 from hydra.core.hydra_config import HydraConfig
 import hydra
@@ -44,7 +44,7 @@ def train_stack(conf: DictConfig) -> None:
    # logging.info(f"{''.join(['#']*100)}")
     selection = 'stacked_'+conf.ts.name+'_'+version
 
-
+    
 
     ######################################################################################################
     
@@ -64,54 +64,67 @@ def train_stack(conf: DictConfig) -> None:
     input_columns = []
     i = 0
     ## Collect all the prediction in the selected set
-    
+    original_data = None
+    predicted_columns = []
     for conf_tmp in files:
+        
         conf_tmp =  OmegaConf.load(conf_tmp) 
-        conf_tmp.inference.set = conf.stack.set
-        conf_tmp.inference.rescaling = conf.stack.rescaling
-        conf_tmp.inference.batch_size = conf.stack.get('batch_size',conf_tmp.inference.batch_size)
-        beauty_string(f'PROCESSING {conf_tmp.model.type}_{conf_tmp.ts.name}_{conf_tmp.ts.version} ','block',VERBOSE)
+        if conf_tmp.stack.dirpath is None:
+            conf_tmp.inference.set = conf.stack.set
+            conf_tmp.inference.rescaling = conf.stack.rescaling
+            conf_tmp.inference.batch_size = conf.stack.get('batch_size',conf_tmp.inference.batch_size)
+            beauty_string(f'PROCESSING {conf_tmp.model.type}_{conf_tmp.ts.name}_{conf_tmp.ts.version} ','block',VERBOSE)
 
 
+            if original_data is None:
+                if conf_tmp.dataset.dataset == 'custom': 
+                    beauty_string('PLEASE WRITE CUSTOM PROCESSING FUNCTION','block',VERBOSE)
+                    ##
+                else:
+                    original_data, columns = read_public_dataset(**conf_tmp.dataset)
+                    targets = ['y']
 
-        try:
-            
-            _,prediction, _ = inference(conf_tmp)
-            
-            ##this can be more informative but the names are too long
-            #prediction['model'] = f'{conf_tmp.model.type}_{conf_tmp.ts.name}_{conf_tmp.ts.version}'
-            model_features = [c for c in prediction.columns if ('pred' in c or 'median' in c)]
-            real_features = [c for c in prediction.columns if not any( k in c for k in ['median','pred','lag','low','high','time']  )]
-            ##renaming columns
-            prediction = prediction[real_features+model_features+['time','lag']]
-            mapping = {}            
-            for j,col in enumerate(model_features):
-                mapping[col] = f'pred_model_{i}_target_{j}'
-            prediction.rename(columns=mapping,inplace=True)
-            input_columns+=list(mapping.values())
-            
-            
-            
-            if predictions is None:
-                predictions = prediction[['time','lag']+list(mapping.values())+real_features]
-                targets = real_features
-                LAG = int(predictions.lag.max())
+            try:
+
+                ##change skip otherwise a lot of non consecutive points
+                conf_tmp.split_params.skip_step = 1 ##force this! and pass to the inference module
+                _,prediction, _ = inference(conf_tmp,conf_tmp.split_params)
+                
+                ##this can be more informative but the names are too long
+                #prediction['model'] = f'{conf_tmp.model.type}_{conf_tmp.ts.name}_{conf_tmp.ts.version}'
+                model_features = [c for c in prediction.columns if ('_pred' in c or '_median' in c)]
+                real_features = [c for c in prediction.columns if not any( k in c for k in ['median','pred','lag','low','high','time']  )]
+                ##renaming columns
+                prediction = prediction[real_features+model_features+['time','lag']]
+                mapping  ={}
+                for j,col in enumerate(model_features):
+                    mapping[col] = f'pred_model_{i}_target_{j}'
+                prediction.rename(columns=mapping,inplace=True)
+                input_columns+=list(mapping.values())
+                predicted_columns+=list(mapping.values())
+
+                
+                if predictions is None:
+                    predictions = prediction[['time','lag']+list(mapping.values())+real_features]
+                    targets = real_features
+                    LAG = int(predictions.lag.max())
 
 
-            else:
-                assert(len(set(model_features).difference(set(model_features)))==0), print('Check models, seems with different targets')
-                prediction = prediction[['time','lag']+list(mapping.values())]
-                predictions = pd.merge(predictions, prediction)
-            N_models+=1
-            models_used.append(conf_tmp)
-            i+=1
-        except Exception as e:
-            import traceback
-            beauty_string(traceback.format_exc(),'',True)
-            beauty_string(f'Can not load model {conf_tmp.model.type}_{conf_tmp.ts.name}_{conf_tmp.ts.version} {e}','block', True)
-    
-    
-    
+                else:
+                    assert(len(set(model_features).difference(set(model_features)))==0), print('Check models, seems with different targets')
+                    prediction = prediction[['time','lag']+list(mapping.values())]
+                    predictions = pd.merge(predictions, prediction)
+                N_models+=1
+                models_used.append(conf_tmp)
+                i+=1
+            except Exception as e:
+                import traceback
+                beauty_string(traceback.format_exc(),'',True)
+                beauty_string(f'Can not load model {conf_tmp.model.type}_{conf_tmp.ts.name}_{conf_tmp.ts.version} {e}','block', True)
+                import pdb
+                pdb.set_trace()
+ 
+
     beauty_string(f'USING {N_models} models','section',VERBOSE)
 
     
@@ -128,23 +141,59 @@ def train_stack(conf: DictConfig) -> None:
     predictions = extend_time_df(predictions,freq,group='lag',global_minmax=True).merge(predictions,how='left')
     predictions['lag_m'] = predictions.lag.values
     silly_model = conf.ts.get('silly',False)
+
+    '''
+    df_long = predictions.drop(columns=['lag_m','prediction_time']).melt(id_vars=["time", "lag"]+targets, var_name="variable", value_name="value")
+    df_long["variable"] = df_long["variable"] + "_lag" + df_long["lag"].astype(str)
+    extra_columns = df_long["variable"].unique()
+    df_wide = df_long.pivot(index=["time"]+targets, columns="variable", values="value").reset_index()
+    df_wide = df_wide.merge(original_data.drop(columns=targets),how='left')
+    
+
+    ts.load_signal(df_wide, 
+                   enrich_cat=conf.ts.enrich,
+                   target_variables=real_features,
+                   past_variables=columns if conf.ts.get('use_covariates',False) else [], 
+                   future_variables=extra_columns,
+                   check_holes_and_duplicates=True,
+                   #cat_past_var=['lag_m'], ##dataset_dependent
+                   #cat_fut_var=['lag_m'],##dataset_dependent
+                   silly_model=silly_model)
+    ts.dataset.sort_values(by=['time'],inplace=True)
+    print(ts)
+
+    import pdb
+    pdb.set_trace()
+
+    model_conf['past_steps'] = LAG
+    model_conf['future_steps'] = LAG
+    model_conf['past_channels'] = len(ts.past_variables)
+    model_conf['future_channels'] = len(ts.future_variables)
+    model_conf['embs_past'] = [ts.dataset[c].nunique() for c in ts.cat_past_var]
+    model_conf['embs_fut'] = [ts.dataset[c].nunique() for c in ts.cat_fut_var]
+    '''
+    
+    predictions = predictions.merge(original_data.drop(columns=targets),how='left')
     ts.load_signal(predictions, 
                    enrich_cat=conf.ts.enrich,
                    target_variables=real_features,
-                   past_variables=[], 
-                   future_variables=input_columns,
+                   past_variables=columns if conf.ts.get('use_covariates',False) else [], 
+                   future_variables=predicted_columns,
                    check_holes_and_duplicates=False,
-                   cat_var=['lag_m'],
+                   cat_past_var=['lag_m'], ##dataset_dependent
+                   cat_fut_var=['lag_m'],##dataset_dependent
                    silly_model=silly_model)
     ts.dataset.sort_values(by=['prediction_time','lag'],inplace=True)
+    print(ts)
 
-    ## TODO qui ci sono delle cose sospette sul futuro...
-    #these parameters depends on the model used not from the config file
+
     model_conf['past_steps'] = LAG
     model_conf['future_steps'] = LAG
-    model_conf['past_channels'] = 1
-    model_conf['future_channels'] = len(targets)*N_models if silly_model is False else len(targets)*N_models + len(targets)
-    model_conf['embs'] = [ts.dataset[c].nunique() for c in ts.cat_var]
+    model_conf['past_channels'] = len(ts.past_variables)
+    model_conf['future_channels'] = len(ts.future_variables)
+    model_conf['embs_past'] = [ts.dataset[c].nunique() for c in ts.cat_past_var]
+    model_conf['embs_fut'] = [ts.dataset[c].nunique() for c in ts.cat_fut_var]    
+    
     model_conf['out_channels'] = len(targets)
 
 
@@ -169,9 +218,10 @@ def train_stack(conf: DictConfig) -> None:
         return 0.0
     
     ##clean folders
-    if  os.path.exists(dirpath):
+    if  (os.path.exists(dirpath)) and (conf.model.get('restart',False) is False):
         shutil.rmtree(dirpath)
-    os.makedirs(dirpath)
+    if  os.path.exists(dirpath) is False:
+        os.makedirs(dirpath)
     conf.train_config.dirpath = dirpath
     ts.set_verbose(VERBOSE)
     ts.set_model(model,config=dict(model_configs=model_conf,
@@ -180,7 +230,7 @@ def train_stack(conf: DictConfig) -> None:
     
     split_params = conf.split_params
     
-    split_params['starting_point'] = {'lag':1}
+    split_params['starting_point'] = {'lag':1} #THIS IS IMPORTANT!
     split_params['past_steps'] = LAG
     split_params['future_steps'] = LAG
